requirements.txt
anthropic==0.8.1
cohere==1.0.0
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
matplotlib==3.8.2
seaborn==0.13.2
python-dotenv==1.0.1
tqdm==4.66.1 